@@ -1,10 +1,5 @@
 type BabelParserOptions = {
-  strictMode?: boolean;
-  looseModules?: boolean;
-  highlightCode?: boolean;
-  nonStandard?: boolean;
+  sourceFilename?: string;
   sourceType?: "module" | "script";
-  filename?: string;
-  features?: Object;
-  plugins?: Object;
+  plugins?: Array<Object>;
 };
