@@ -116,7 +116,7 @@ export function shouldIgnore(
   ignore: Array<RegExp | Function> = [],
   only?: Array<RegExp | Function>,
 ): boolean {
-  filename = slash(filename);
+  filename = filename.replace(/\\/g, "/");
 
   if (only) {
     for (let pattern of only) {
