@@ -1,6 +1,8 @@
 var t = function (f) {
+  var _ref;
+
   arguments.length <= 1 ? undefined : arguments[1];
-  arguments.length <= (arguments.length <= 1 ? 0 : arguments.length - 1) - 1 + 1 ? undefined : arguments[(arguments.length <= 1 ? 0 : arguments.length - 1) - 1 + 1];
+  _ref = (arguments.length <= 1 ? 0 : arguments.length - 1) - 1 + 1, arguments.length <= _ref ? undefined : arguments[_ref];
 };
 
 function t(f) {
