@@ -6,9 +6,12 @@ export default function () {
       ObjectMethod(path) {
         let { node } = path;
         if (node.kind === "method") {
+          const func = t.functionExpression(null, node.params, node.body, node.generator, node.async);
+          func.returnType = node.returnType;
+
           path.replaceWith(t.objectProperty(
             node.key,
-            t.functionExpression(null, node.params, node.body, node.generator, node.async),
+            func,
             node.computed
           ));
         }
