@@ -45,7 +45,7 @@ defineType("ClassImplements", {
 defineType("ClassProperty", {
   visitor: ["key", "value", "typeAnnotation", "decorators"],
   builder: ["key", "value", "typeAnnotation", "decorators", "computed"],
-  aliases: ["Flow", "Property"],
+  aliases: ["Property"],
   fields: {
     computed: {
       validate: assertValueType("boolean"),
